requirements.txt
Flask==3.1.0
openai==1.58.1
jsonschema==4.23.0
python-dotenv==1.0.1
